--- src/KuniCharacter.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "AUI/Common/AString.h"
-
-namespace kuni_character
-{
-    AString getBasePrompt();
-    AString getAppearancePrompt();
-}
