+++ src/tools/remove_and_ban_chat.cpp
@@ -39,7 +39,7 @@ OpenAITools::Tool tools::removeAndBanChat(_<ITelegramClient> telegram) {
         .handler = [telegram = std::move(telegram)](OpenAITools::Ctx ctx) -> AFuture<AString> {
             auto chatId = util::jsonAsLongInt(ctx.args["chat_id"]).valueOrException("chat_id integer is required");
             ALogger::info("remove_and_ban_chat") << "remove_and_ban_chat: chat_id" << chatId;
-            if (chatId == config::PAPIK_CHAT_ID) {
+            if (chatId == config().papikChatId) {
                 // precaution -- can't delete chat with papik
                 co_return "Failed";
             }
