+++ src/tools/send_telegram_message.cpp
@@ -4,6 +4,7 @@
 
 #include "send_telegram_message.h"
 
+#include "prompts.h"
 #include "AUI/IO/AFileInputStream.h"
 #include "AUI/Util/kAUI.h"
 #include "llmui/telegram.h"
@@ -129,7 +130,7 @@ OpenAITools::Tool tools::sendTelegramMessage(
                         "- ахахаххаа\n"
                         "- ты смешной\n"
                         "- научишь также?\n";
-                    if constexpr (config::CAPABILITY_USE_STICKERS) {
+                    if (config().capabilityUseStickers) {
                         msg += "Alternatively, you can use a sticker (#sticker_send)\n";
                     }
                     co_return msg;
@@ -138,20 +139,20 @@ OpenAITools::Tool tools::sendTelegramMessage(
 
 #ifndef AUI_TESTS_MODULE
             if (photoFilename.empty() && audioFilename.empty()) {
-                bool shouldRemind = std::uniform_real_distribution<>(0.0, 1.0)(gRandomEngine) < config::TOOL_REMINDER_CHANCE;
+                bool shouldRemind = std::uniform_real_distribution<>(0.0, 1.0)(gRandomEngine) < config().toolReminderProbability;
                 if (shouldRemind) {
                     bool usePhoto = std::uniform_real_distribution<>(0.0, 1.0)(gRandomEngine) < 0.5;
                     AString reminderMessage = "Constant texting is too dull for the user!\n";
 
-                    if constexpr (config::CAPABILITY_USE_STICKERS) {
+                    if (config().capabilityUseStickers) {
                         reminderMessage += "- Consider using stickers! #sticker_send\n";
                     }
 
-                    if (usePhoto && config::CAPABILITY_TAKE_PHOTO) {
+                    if (usePhoto && config().capabilityTakePhoto) {
                         reminderMessage += "- Consider sending photos from your gallery or generated by #take_photo tool to make the conversation more lively and engaging!\n";
                         throw AException(reminderMessage);
                     }
-                    if (!usePhoto && config::CAPABILITY_RECORD_AUDIO) {
+                    if (!usePhoto && config().capabilityRecordVoice) {
                         reminderMessage += "- Consider recording voice notes by #record_audio tool and sending them to make the conversation more lively and engaging!\n";
                         throw AException(reminderMessage);
                     }
@@ -209,7 +210,7 @@ OpenAITools::Tool tools::sendTelegramMessage(
             // dirty fix: skip similarity checks if a photo was attached: llm's comment on photo is not much
             // important
             if (!message.empty() && photoFilename.empty()) {
-                auto target = co_await openAI->embedding({ .config = config::ENDPOINT_EMBEDDING }, message);
+                auto target = co_await openAI->embedding({ .config = config().embedding }, message);
                 static AMap<AString, std::valarray<double>> embeddings;
                 double maxSimilarity = 0.0;
                 double avgSimilarity = 0.0;
@@ -227,12 +228,12 @@ OpenAITools::Tool tools::sendTelegramMessage(
                     auto text = llmui::extractMessageTypeAndText(*i);
                     auto& embedding = embeddings[text];
                     if (embedding.size() != target.size()) {
-                        embedding = co_await openAI->embedding({ .config = config::ENDPOINT_EMBEDDING }, text);
+                        embedding = co_await openAI->embedding({ .config = config().embedding }, text);
                     }
                     const auto similiarity = util::cosine_similarity(target, embedding);
                     avgSimilarity += similiarity;
                     maxSimilarity = std::max(maxSimilarity, similiarity);
-                    if (similiarity > config::REPEAT_YOURSELF_TRIGGER_MAX + giveAHeadStart) {
+                    if (similiarity > config().antiRepeatTriggerMax + giveAHeadStart) {
                         giveAHeadStart += 0.07; // relax repeating after itself check
                         ALogger::warn(LOG_TAG) << "LLM is repeating itself: (maxSimilarity=" << maxSimilarity << ")" << message;
                         // If LLM generates a follow-up that repeats meaning of its previous responses,
@@ -251,11 +252,11 @@ OpenAITools::Tool tools::sendTelegramMessage(
                         // I'm trying to make Kuni more lazy by suggesting closing a chat on a low-quality
                         // follow-up.
 
-                        throw AException(config::REPEAT_YOURSELF_PROMPT);
+                        throw AException(prompts().antiRepeatPrompt);
                     }
                 }
                 avgSimilarity /= countOfKunisMessages;
-                if (avgSimilarity > config::REPEAT_YOURSELF_TRIGGER_AVG + giveAHeadStart) {
+                if (avgSimilarity > config().antiRepeatTriggerAvg + giveAHeadStart) {
                     giveAHeadStart += 0.07; // relax repeating after itself check
                     // LLM figured out threshold of REPEAT_YOURSELF_TRIGGER_MAX and indeed it generates
                     // slightly more variative responses, but their general direction and structure feels
@@ -273,13 +274,13 @@ OpenAITools::Tool tools::sendTelegramMessage(
                     // switch topic
 
                     ALogger::warn(LOG_TAG) << "LLM is repeating itself: (avgSimilarity=" << avgSimilarity << ")" << message;
-                    throw AException(config::REPEAT_YOURSELF_PROMPT);
+                    throw AException(prompts().antiRepeatPrompt);
                 }
 
                 giveAHeadStart = 0.0; // reset indulgence
 
-                if (embeddings.size() >= config::REPEAT_YOURSELF_MAX_HISTORY) {
-                    ALOG_DEBUG(LOG_TAG) << "Dropped \"repeat yourself\" history";
+                if (embeddings.size() >= config().antiRepeatMaxHistory) {
+                    ALOG_DEBUG(LOG_TAG) << "Dropped \"anti repeat yourself\" history";
                     embeddings.clear();
                 }
                 ALOG_DEBUG(LOG_TAG) << "\"repeat yourself\" maxSimilarity=" << maxSimilarity << " avgSimilarity=" << avgSimilarity;
