+++ src/util/is_accessible_from_lockdown.cpp
@@ -6,14 +6,14 @@
 
 AFuture<bool> util::isAccessibleFromLockdown(ITelegramClient& telegram, int64_t chatId) {
 #ifdef AUI_TESTS_MODULE
-    co_return config::PAPIK_CHAT_ID == chatId;
+    co_return config().papikChatId == chatId;
 #else
-    switch (config::LOCKDOWN_MODE) {
-        case config::LockdownMode::NONE: {
+    switch (config().lockdown) {
+        case Config::LockdownMode::NONE: {
             co_return true;
         }
 
-        case config::LockdownMode::CONTACTS_ONLY: {
+        case Config::LockdownMode::CONTACTS_ONLY: {
             static auto contacts = co_await telegram.sendQueryWithResult(ITelegramClient::toPtr(td::td_api::getContacts()));
             if (ranges::contains(contacts->user_ids_, chatId)) {
                 co_return true;
@@ -21,8 +21,8 @@ AFuture<bool> util::isAccessibleFromLockdown(ITelegramClient& telegram, int64_t 
             [[fallthrough]];
         }
 
-        case config::LockdownMode::PAPIK_ONLY: {
-            if (chatId == config::PAPIK_CHAT_ID) {
+        case Config::LockdownMode::PAPIK_ONLY: {
+            if (chatId == config().papikChatId) {
                 co_return true;
             }
             [[fallthrough]];
